main: Pass args to store def init funcs which require them
Refactoring in main to pass the number of required parameters to the
store definitions which export an initialization function which requires
some parameters.

The passed parameters are fake objects which can be called or accessed
to any property, because it's an instance of recursive proxy. Actually,
it's the same instance than the one used for 'window' and 'document'
globals.

These changes make the tool to work with those store definitions which
require some arguments (dependencies) which are used at initialization.

This commit makes to pass all the tests again.